refactor: 承認系 value object を domain/values/approval/ へ区画化 (#421)
承認系 VO 13個と隣接テストを domain/values/approval/ 配下へ集約。
import パスは values/<VO> → values/approval/<VO> の単一変換（エイリアス・
相対の両形式）。移動VO同士の相対import（StepRejection→./RejectionComment）
とテストの自参照は共移動のため不変。

テスト配置は approval/__tests__/（隣接維持）。build/eslint(estimate)/test green。
