+++ src/server/subdomains/estimate/application/shared/__tests__/resolveApprovalGoalTiersByDepth.test.ts
@@ -1,5 +1,5 @@
 import { BusinessRuleViolationError } from "@server/shared/errors/DomainError";
-import { ApprovalGoalTier } from "@subdomains/estimate/domain/values/ApprovalGoalTier";
+import { ApprovalGoalTier } from "@subdomains/estimate/domain/values/approval/ApprovalGoalTier";
 import { PositionId } from "@subdomains/position/domain/values/PositionId";
 import { describe, expect, it } from "vitest";
 import {
