+++ src/server/subdomains/estimate/domain/entities/EstimateApprovalExemption.ts
@@ -1,6 +1,6 @@
 import { EmployeeId } from "@subdomains/employee/domain/values/EmployeeId";
-import { EstimateApprovalExemptionId } from "../values/EstimateApprovalExemptionId";
-import { EstimateExemptionReason } from "../values/EstimateExemptionReason";
+import { EstimateApprovalExemptionId } from "../values/approval/EstimateApprovalExemptionId";
+import { EstimateExemptionReason } from "../values/approval/EstimateExemptionReason";
 import { EstimateVariationId } from "../values/EstimateVariationId";
 
 /**
