+++ src/server/subdomains/estimate/domain/entities/__tests__/AfterRepairEstimateDetail.test.ts
@@ -1,7 +1,7 @@
 import { ProductId } from "@subdomains/product/domain/values/ProductId";
 import { describe, expect, it } from "vitest";
 import { AfterRepairEstimateDetailId } from "../../values/AfterRepairEstimateDetailId";
-import { EmergencyReason } from "../../values/EmergencyReason";
+import { EmergencyReason } from "../../values/approval/EmergencyReason";
 import { FaultDescription } from "../../values/FaultDescription";
 import { AfterRepairEstimateDetail } from "../AfterRepairEstimateDetail";
