+++ src/server/subdomains/estimate/domain/policies/ApprovalRequirementPolicy.ts
@@ -1,6 +1,6 @@
 import { ProductCategory } from "@subdomains/product/domain/values/ProductCategory";
-import { ApprovalGoalTier } from "../values/ApprovalGoalTier";
-import { EstimateExemptionReason } from "../values/EstimateExemptionReason";
+import { ApprovalGoalTier } from "../values/approval/ApprovalGoalTier";
+import { EstimateExemptionReason } from "../values/approval/EstimateExemptionReason";
 import { EstimateType } from "../values/EstimateType";
 import { Money } from "../values/Money";
