+++ src/server/subdomains/estimate/domain/values/approval/ApplicationStatus.ts
